:hp-cover: Attribute in Readme.adoc is incorrect.
Hi @anthonny 

According to the docs, this is how you declare a blog post cover image:

```
= Build API Docs for the RHQ Project 
:hp-tags: Maven, RHQ, API Docs
:hp-image: http://github.com/jaredmorgs/jaredmorgs.github.io/images/covers/open_source.jpg
:published_at: 2015-02-16
```

Based on other images I've used in my blogs, you need to declare this as:

:hp-image: http://jaredmorgs.github.io/images/covers/open_source.jpg

This works for me on jaredmorgs.github.com. 

I've created this PR to fix the issue.

I also cleaned up the other parameter descriptions.