[Review Required] Added two content: blocks
Hi @anthonny. I wanted to layer the "settings saved" message with another message that tells the user not to make further configuration changes until HubPress confirms posts were successfully regenerated. 

Please verify the way I've done this (using a content: block on each line with a different message) can be parsed by the code correctly. If not, please run the sentences together in the same content: block (and sorry for the trouble).